com.unity.asset-manager-for-unity@0.7.0 ## [0.7.0] - 2024-05-17
[Added]
* A new preprocessing directive "AM4U_ENABLE_ASSERTIONS" that enable assertion when set.

[Changed]
* Upload mode set to Override by default
* Show In Project will ping the first file instead of the parent folder
* Draft assets will display their version in yellow like on the web app
* Disabling Remove button while re-import is in progress
* Remove the ImportOperation.ImportAsync IAsset parameter, which was a duplicate of the IAssetData given in the constructor
* Tracking recognizes version changes and out-of-date assets will be updated to the latest version on import
* Cache location is now displayed with an absolute path in the user settings
* Removed download progress indicators from the import operation
* Upload progress indicators are no longer stick
* Improved offline mode and domain reload support

[Fixed]
* Import to action wasn't working
* Don't show import progression of -100% anymore
* Imported asset without .meta file are now properly re-uploaded (need to be imported first)
* Re-uploading a new version if an asset was not tracked properly
* Upload progress not displaying error if details panel is opened
* Normalizing path's separators in AM4U settings
* Fixed Upload files displaying Assets/ folder
* Fixed critical issue where Role and Permissions where not serialized properly
* Fixed error when cancelling Import To folder selection
* Fixed error when selecting Assets folder using Import To dialog
* Removing unused constant causing warning
* Improved search to better support underscores and other separator
* Fixed missing Reveal In Finder label in settings
* Asset files occasionally failing to re-import and being deleted.
* Fixed the import of files with no extensions
* Wrong search results when adding search words while the grid view has not finished loading.
* Fixed a bug where upload progress indicators would be duplicated.

+++ Editor/Services/IOProxy.cs
@@ -19,7 +19,6 @@ namespace Unity.AssetManager.Editor
         string FileReadAllText(string filePath);
         void FileWriteAllText(string filePath, string text);
         IEnumerable<string> EnumerateFiles(string path, string searchPattern, SearchOption searchOption);
-        string GetRelativePathToProjectFolder(string path);
         string GetUniqueTempPathInProject();
         bool DeleteAllFilesAndFoldersFromDirectory(string path);
         FileStream Create(string path, int bufferSize, FileOptions options) => File.Create(path, bufferSize, options);
@@ -27,11 +26,6 @@ namespace Unity.AssetManager.Editor
 
     class IOProxy : BaseService<IIOProxy>, IIOProxy
     {
-        string m_FullAssetsFolderPath;
-
-        string FullAssetsFolderPath =>
-            m_FullAssetsFolderPath ??= Path.GetFullPath(Path.Combine(Application.dataPath, "../"));
-
         public bool FileExists(string filePath)
         {
             return File.Exists(filePath);
@@ -39,6 +33,9 @@ namespace Unity.AssetManager.Editor
 
         public void FileMove(string sourceFilePath, string destinationFilePath)
         {
+            if (!FileExists(sourceFilePath))
+                return;
+
             new FileInfo(destinationFilePath).Directory?.Create();
             File.Move(sourceFilePath, destinationFilePath);
         }
@@ -80,15 +77,6 @@ namespace Unity.AssetManager.Editor
         public void FileWriteAllText(string filePath, string text) => File.WriteAllText(filePath, text);
 
         public IEnumerable<string> EnumerateFiles(string path, string searchPattern, SearchOption searchOption) => Directory.EnumerateFiles(path, searchPattern, searchOption);
-
-        public string GetRelativePathToProjectFolder(string path)
-        {
-            // It was seen that when using this in conjunction with some Unity APIs (e.g. SceneManager.GetScenePath)
-            // it was not working because of backslashes, so we normalize it to forward instead here
-            return path.StartsWith(FullAssetsFolderPath) ?
-                path.Substring(FullAssetsFolderPath.Length).Replace(Path.DirectorySeparatorChar, '/') :
-                path;
-        }
 
         public string GetUniqueTempPathInProject() => FileUtil.GetUniqueTempPathInProject();
